Mail: Fix wordwrap for system mails (Mantis 33198)

+++ Services/Mail/classes/BackgroundTask/class.ilMailDeliveryJob.php
@@ -32,7 +32,7 @@ class ilMailDeliveryJob extends AbstractJob
             json_encode(array_slice($arguments, 0, 5), JSON_THROW_ON_ERROR | JSON_PRETTY_PRINT)
         ));
 
-        if ((int) $input[0]->getValue() === ANONYMOUS_USER_ID) {
+        if ((int) $input[0]->getValue() === (int) ANONYMOUS_USER_ID) {
             $mail = new ilMail((int) $input[0]->getValue());
         } else {
             $mail = new ilFormatMail((int) $input[0]->getValue());
